Fix(Dashboard)Added new download buttons and separated download component
